+++ storybook/cypress/e2e/button-group.cy.ts
@@ -1,14 +1,21 @@
 describe("Button Group", () => {
-  beforeEach(() => {
-    cy.visit("http://localhost:45000/iframe.html?id=core-button-group--default");
-  });
-
   it("is accessible", () => {
+    cy.visit("http://localhost:45000/iframe.html?id=core-button-group--default");
     cy.injectAxe();
     cy.dsoCheckA11y("dso-button-group.hydrated");
   });
 
-  it("matches snapshot", () => {
-    cy.get("dso-button-group.hydrated").matchImageSnapshot();
-  });
+  const directions = ["row", "column"];
+  const buttonElements = ["button", "anchor", "icon-button"];
+
+  for (const direction of directions) {
+    for (const buttonElement of buttonElements) {
+      it(`matches snapshot ${direction} of ${buttonElement}`, () => {
+        cy.visit(
+          `http://localhost:45000/iframe.html?id=core-button-group--default&args=buttonElement:${buttonElement}`,
+        );
+        cy.get("dso-button-group.hydrated").invoke("prop", direction).matchImageSnapshot();
+      });
+    }
+  }
 });