+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Changed
 * **BREAKING** Document Card: Label positie aanpassen ([#3347](https://github.com/dso-toolkit/dso-toolkit/issues/3347))
+* Map Controls: Icon Button integratie ([#3362](https://github.com/dso-toolkit/dso-toolkit/issues/3362))
 
 ### Task
 * Angular Storybook: Vervang `*ngIf` door `@if` block ([#3304](https://github.com/dso-toolkit/dso-toolkit/issues/3304))